Reorganize smoke tests.
